Functions from past projects
This repository is for my own possible reference in the future, and to show some of my python codes.
If the functions happen to be useful for your assignments (if you are having the class I have taken before), please do NOT simply copy and paste. It may get we both in trouble.

All functions in hw2 are included in other files.
hw6 is damaged. I will try to recover it some other day.